+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## [Unreleased]
 ### Changed
-- Client version updated on [5.4.2](https://github.com/reportportal/client-java/releases/tag/5.4.2), by @HardNorth
+- Client version updated on [5.4.3](https://github.com/reportportal/client-java/releases/tag/5.4.3), by @HardNorth
 - Replace "jsr305" annotations with "jakarta.annotation-api", by @HardNorth
 - Switch on use of `Instant` class instead of `Date` to get more timestamp precision, by @HardNorth