+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Added
 - TestNG retries handling, by @HardNorth
 ### Changed
-- Client version updated to [5.4.8](https://github.com/reportportal/client-java/releases/tag/5.4.8), by @HardNorth
+- Client version updated to [5.4.9](https://github.com/reportportal/client-java/releases/tag/5.4.9), by @HardNorth
 
 ## [5.4.3]
 ### Changed